Release cargo command made.
Changes to have obtain and release cargo work with both tilt and grab subsystems.

+++ src/main/cpp/subsystems/TiltSystem.cpp
@@ -25,7 +25,7 @@ void TiltSystem::off() {
   
 }
 
-void TiltSystem::Tilt(int direction) {
+void TiltSystem::Tilt(double direction) {
   
 }
